[Runners] Raise the start event before the tests executes. (#59)
The OnTestStarted was being called in the HandleAfterTestStarting. That
means that the event might not be raised if there is an issue with the
test setup, making it hard to track. Moved the call to the
HandleBeforeTestStarting to notify the test is going to strat which will
execute before any issue occurs.

fixes: https://github.com/dotnet/xharness/issues/54