[iOS] Ensure that we do generate xml when we call the CLI.
During the refactor we removed the env vars that are passed to the
runners to decide if xml is used and the jargon to use. The env
paramenters have been added and tests updated. Now the runners will get
notified that xml is used (always) and the jargon (default to xunit).

fixes: https://github.com/dotnet/xharness/issues/98
fixes: https://github.com/dotnet/xharness/issues/77