[Runners] Do not filter per Attr, make client use traits.
We support traits from the very first release, those have to be passed in
a file that is provided by the child of the entry point:

IgnoreFilesDirectory.

That directory can contain two diff files:

* xunit-excludes.txt
* nunit-excludes.txt

I have updated the docs to describe what should be done since is the
root of the misscommunication.

fixes: https://github.com/dotnet/xharness/issues/137