[Runners] Workaround to discover all the active issues.
The ActiveIssue discoverer does have a bug in which it does not add the
traits on Android or iOS. The fix was landed here:
https://github.com/dotnet/arcade/pull/5425 but this commit adds a
workaround. Issue https://github.com/dotnet/xharness/issues/156 to
track the workaround and make sure it is removed.

fixes: https://github.com/dotnet/xharness/issues/53

PS: This time it really fixes the issue ;)