Adds support for skipping tests when SkipTestException is raised
If a test fails and the exception type is SkipTestException, then the runner
will flag that as a skipped test.  This is to support tests that depend on something in the xunit
pipeline that may not be supported on mobile.  We don't want to have to go into every test and
mark ConditionalFact/Theory.