[iOS] Ensure we do select devices according to their architecture.
The FindDevice method returns the first device found, but will not
filter over the architecture. Move to use LoadDevices and then use the
collections to use the first device that matches the desired arch.

With this commit, xharness will pick the **correct** device if two different
devices are present in the host.

fixes: https://github.com/dotnet/xharness/issues/86