[iOS] Allow users to use lldb to debug issues in test applications.
Add support for the lldb feature in mtouch. User can pass a new flag to
state that lldb should be enabled, the CLI will then provie a command to
copy paste to access the lldb session.

The aggregated log was created to state which of the ILogs that was
passed should be considered the default log and therefore the one to
return the FullPath and the StreamReader.