[iOS] Fix the mess we had with the logs.
The initial implementation of the logs, yet usefull, it was wrong. It
made it very easy to pass Logs that do have certain missing properties
and methods to other classes that used those methods, resulting in
runtime issues.

We move to have different interfaces that will allow to say which is the
minimun amonunt of methods to be implemented (ILog) or the maximun one
(IFileBackedLog) so that the compiler will help us to not pass a
MemoryLog or a ConsoleLog to those methods that will like to use, for
example, the FullPath of the log.

This might look like a breaking change for xamarin-macios, but whatever
it breaks was a hidden bug.