Improve wasm test logging
Don't show PASS/SKIP messages by default in the console log, unless verbose logging is enabled in xharness.
We capture all output and store it in the wasm-console.log file in the xharness output directory instead.

Also add special handling for FAIL messages so that they show up as a single entry in the console log output.