[wasm] test-browser: Add support for Edge
Selenium's EdgeDriver seems to be looking for `MicrosoftWebDriver.exe`,
but the driver executable is called `msedgedriver` on mac/linux/windows.
I couldn't find a way to create the service by just specifying the
driver name, so instead we try to find the file in `$PATH`, and use that
to start the driver service.

The code for Edge, and Chrome is pretty much the same, except for the
concrete types, so sharing all that.