Add script for running E2E tests (#585)
This script is a quick way to run the XHarness E2E tests. These test are located in tests/integration-tests and require the Arcade and Helix SDK. To run them, you need to invoke these through MSBuild which makes the process a bit cumbersome. This script should make things easier.