[wasm] Add scanning test output for error patterns
In some cases, the runtime might crash, for example, due to an
assertion. But the app continues to execute, causing v8/chrome to exit
with the expected exit code. This causes xharness to treat the tests as
passing, even though the runtime had crashed.

To catch such cases, this adds support for scanning the output for some
error patterns, and adds a new argument for `wasm test`, and `wasm
test-browser`:

    `--error-patterns=<file>`

This expects a text file with one pattern per line. The first character
of the line is treated special:

    '@': the rest of the line is a substring to be used as the pattern
         (case insensitive)
    '%': the rest of the line is a .net regex string
    '#': ignore the line completely

Example patterns file:
```
%console.error: Unhandled *[Ee]xception

@== JS stack trace
```

- this ignores whitespace only lines.
- And stops scanning after the first match