Detect app launch timeout on Apple (#786)
The shared Xamarin code allowed for a short window in-between the app being installed and launched but not having connected over TCP to XHarness' listener.
We add a more granular code that will describe this case so that we don't need to return `TIMED_OUT` when `--launch-timeout` occurs since `TIMED_OUT` means that tests themselves are running too long. This new launch timeout is then categorized as infra failure rather than user's fault.