Get Android device architecture for diagnostic data (#795)
Currently, we do not include device's architecture in diagnostic data but rather the first of a list of all supported archs by the device. This PR adds a new ADB call that queries for the actual architecture of the device.

Secondly, this PR brings the Android and Apple state commands on-par and allows to export Android state information as JSON too (useful for querying devices over Helix for example).