Correctly wait for completion of DelegatingXmlCreationSink
With https://github.com/dotnet/runtime/pull/62501 we were only waiting for
completion of the DelegatingExecutionSummarySink, not the DelegatingXmlCreationSink.

Since DelegatingXmlCreationSink doesn't provide a callback when testing is done we
need to add a wrapper execution sink that listens for the ITestAssemblyFinished message.

Fixes https://github.com/dotnet/xharness/issues/903